Rename Button component to polygonButton
Renamed the Button component file to polygonButton.tsx and updated its import in the about page to reflect the new name.
