+++ app/about/page.tsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import Button from '../components/ui/Button'
+import Button from '../components/ui/polygonButton'
 
 const page = () => {
   return (
